Condense test files
